requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiohttp==3.8.6
pendulum==2.1.2
Pillow==10.0.1
python-dotenv==1.0.0
vkbottle==4.3.12
